Homework2/FAAS/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.10.0